Fix out of sync view changes
It can happen that a conode gets stuck in the middle of a view change
waiting for more requests and the cothority adds one or more new blocks.
This fixes the situation by first ignoring invalid requests and by
cleaning previous states when a new block is propagated.

Fixes #1583
