+++ byzcoin/service_test.go
@@ -5,7 +5,6 @@ import (
 	"crypto/sha256"
 	"encoding/binary"
 	"errors"
-	"math"
 	"strings"
 	"testing"
 	"time"
@@ -13,7 +12,6 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"go.dedis.ch/cothority/v3"
-	"go.dedis.ch/cothority/v3/byzcoin/viewchange"
 	"go.dedis.ch/cothority/v3/darc"
 	"go.dedis.ch/cothority/v3/darc/expression"
 	"go.dedis.ch/cothority/v3/skipchain"
@@ -1832,143 +1830,6 @@ func TestService_SetBadConfig(t *testing.T) {
 		if badConfig.Roster.List[0].Equal(config.Roster.List[0]) {
 			require.Fail(t, "found a bad config")
 		}
-	}
-}
-
-// TestService_ViewChange is an end-to-end test for view-change. We kill the
-// first nFailures nodes, where the nodes at index 0 is the current leader. The
-// node at index nFailures should become the new leader. Then, we try to send a
-// transaction to a follower, at index nFailures+1. The new leader (at index
-// nFailures) should poll for new transactions and eventually make a new block
-// containing that transaction. The new transaction should be stored on all
-// followers. Finally, we bring the failed nodes back up and they should
-// contain the transactions that they missed.
-func TestService_ViewChange(t *testing.T) {
-	testViewChange(t, 4, 1, 4*time.Second)
-}
-
-func TestService_ViewChange2(t *testing.T) {
-	if testing.Short() {
-		t.Skip("doesn't work on travis correctly due to byzcoinx timeout issue, see #1428")
-	}
-	testViewChange(t, 7, 2, 4*time.Second)
-}
-
-func testViewChange(t *testing.T, nHosts, nFailures int, interval time.Duration) {
-	rw := rotationWindow
-	defer func() {
-		rotationWindow = rw
-	}()
-	rotationWindow = 3
-	s := newSerN(t, 1, interval, nHosts, true)
-	defer s.local.CloseAll()
-
-	for _, service := range s.services {
-		service.SetPropagationTimeout(2 * interval)
-	}
-
-	// Wait for all the genesis config to be written on all nodes.
-	genesisInstanceID := InstanceID{}
-	for i := range s.services {
-		s.waitProofWithIdx(t, genesisInstanceID.Slice(), i)
-	}
-
-	// Stop the first nFailures hosts then the node at index nFailures
-	// should take over.
-	for i := 0; i < nFailures; i++ {
-		log.Lvl1("stopping node at index", i)
-		s.services[i].TestClose()
-		s.hosts[i].Pause()
-	}
-	// Wait for proof that the new expected leader, s.services[nFailures],
-	// has taken over. First, we sleep for the duration that an honest node
-	// will wait before starting a view-change. Then, we sleep a little
-	// longer for the view-change transaction to be stored in the block.
-	for i := 0; i < nFailures; i++ {
-		time.Sleep(time.Duration(math.Pow(2, float64(i))) * s.interval * rotationWindow)
-	}
-	time.Sleep(2 * s.interval)
-	config, err := s.services[nFailures].LoadConfig(s.genesis.SkipChainID())
-	require.NoError(t, err)
-	log.Lvl2("Verifying roster", config.Roster.List)
-	require.True(t, config.Roster.List[0].Equal(s.services[nFailures].ServerIdentity()))
-
-	// check that the leader is updated for all nodes
-	for _, service := range s.services[nFailures:] {
-		// everyone should have the same leader after the genesis block is stored
-		leader, err := service.getLeader(s.genesis.SkipChainID())
-		require.NoError(t, err)
-		require.NotNil(t, leader)
-		require.True(t, leader.Equal(s.services[nFailures].ServerIdentity()))
-	}
-
-	// try to send a transaction to the node on index nFailures+1, which is
-	// a follower (not the new leader)
-	tx1, err := createOneClientTx(s.darc.GetBaseID(), dummyContract, s.value, s.signer)
-	require.NoError(t, err)
-	s.sendTxTo(t, tx1, nFailures+1)
-
-	// wait for the transaction to be stored on the new leader, because it
-	// polls for new transactions
-	pr := s.waitProofWithIdx(t, tx1.Instructions[0].InstanceID.Slice(), nFailures)
-	require.True(t, pr.InclusionProof.Match(tx1.Instructions[0].InstanceID.Slice()))
-
-	// The transaction should also be stored on followers
-	for i := nFailures + 1; i < nHosts; i++ {
-		pr = s.waitProofWithIdx(t, tx1.Instructions[0].InstanceID.Slice(), i)
-		require.True(t, pr.InclusionProof.Match(tx1.Instructions[0].InstanceID.Slice()))
-	}
-
-	// We need to bring the failed (the first nFailures) nodes back up and
-	// check that they can synchronise to the latest state.
-	for i := 0; i < nFailures; i++ {
-		log.Lvl1("starting node at index", i)
-		s.hosts[i].Unpause()
-		require.NoError(t, s.services[i].startAllChains())
-	}
-	for i := 0; i < nFailures; i++ {
-		pr = s.waitProofWithIdx(t, tx1.Instructions[0].InstanceID.Slice(), i)
-		require.True(t, pr.InclusionProof.Match(tx1.Instructions[0].InstanceID.Slice()))
-	}
-
-	log.Lvl1("Sending 1st tx")
-	tx1, err = createOneClientTxWithCounter(s.darc.GetBaseID(), dummyContract, s.value, s.signer, 1)
-	require.NoError(t, err)
-	s.sendTxToAndWait(t, tx1, nFailures, 10)
-	log.Lvl1("Sending 2nd tx")
-	tx1, err = createOneClientTxWithCounter(s.darc.GetBaseID(), dummyContract, s.value, s.signer, 2)
-	require.NoError(t, err)
-	s.sendTxToAndWait(t, tx1, nFailures, 10)
-	log.Lvl1("Sent two tx")
-}
-
-// Tests that a view change can happen when the leader index is out of bound
-func TestService_ViewChangeForced(t *testing.T) {
-	s := newSerN(t, 1, time.Second, 5, true)
-	defer s.local.CloseAll()
-
-	err := s.services[0].sendViewChangeReq(viewchange.View{LeaderIndex: -1})
-	require.Error(t, err)
-	require.Equal(t, "leader index must be positive", err.Error())
-
-	for i := 0; i < 5; i++ {
-		err := s.services[i].sendViewChangeReq(viewchange.View{
-			ID:          s.genesis.SkipChainID(),
-			Gen:         s.genesis.SkipChainID(),
-			LeaderIndex: 7,
-		})
-		require.NoError(t, err)
-	}
-
-	// more than enough for a view change to complete
-	time.Sleep(10 * time.Second)
-
-	for _, service := range s.services {
-		// everyone should have the same leader after the genesis block is stored
-		leader, err := service.getLeader(s.genesis.SkipChainID())
-		require.NoError(t, err)
-		require.NotNil(t, leader)
-		require.True(t, leader.Equal(s.services[2].ServerIdentity()))
 	}
 }
